requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albumentations==0.5.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
dnspython==1.16.0
easydict==1.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Faker==37.6.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hydra-core==1.1.0
idna==3.7
imageio==2.37.0
img2pdf==0.6.1
imgaug==0.4.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
kiwisolver==1.4.9
kornia==0.5.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
namex==0.1.0
narwhals==2.5.0
networkx==3.5
numpy==2.2.6
omegaconf==2.1.2
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pi_heif==1.1.0
pikepdf==9.11.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
polars==1.33.1
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==3.12.0
pyparsing==3.2.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==1.2.9
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
roboflow==1.2.9
rpds-py==0.27.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.3
streamlit==1.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.9.9
toml==0.10.2
torchmetrics==0.2.0
tornado==6.5.2
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.199
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webdataset==1.0.2
Werkzeug==3.1.3
wheel==0.45.1
wldhx.yadisk-direct==0.0.6
wrapt==1.17.3
yarl==1.20.1
git+https://github.com/advimman/lama.git
fastapi
uvicorn
pymongo
python-dotenv
streamlit
pytesseract
img2pdf
faker
pillow
opencv-python
ultralytics
spacy
numpy